PasswordChange and PasswordForget
Password forget and change uses a form to submit the information needed by the firebase authentication API to reset the password. A class method onSubmit ensures the infor is sent to the API. It also resets the form's input field on a successful request, and shows an error on an erroneous request.The form is also validated before it's submitted.